+++ src/shared/isLogin.js
@@ -1,0 +1,5 @@
+import { getCookie } from "./Cookie";
+
+const isLogin = () => !!getCookie("accessToken");
+
+export default isLogin;
